imany changes

+++ subs/edit.php
@@ -106,22 +106,28 @@ if ($delete && $id) {
             echo $renderer->footer();
             die;
         case 'school':
-            if ($confirm and confirm_sesskey()) {
-                $result=$DB->delete_records(constants::M_TABLE_SCHOOLS,array('id'=>$id));
-                redirect($returnurl);
+
+            //get the school
+            $school = $DB->get_record(constants::M_TABLE_SCHOOLS, array('id' => $id));
+            if(!$school){
+                redirect($returnurl,get_string('badschool',constants::M_COMP),
+                    3,\core\output\notification::NOTIFY_WARNING);
             }
 
-            //cancel the deletion request if their exist subs using this plan
+            //cancel the deletion request if there exist subs using this plan
             $subs = common::fetch_subs_by_school($id);
             if($subs && count($subs)){
                 redirect($returnurl,get_string('existingsubsforschool',constants::M_COMP),
                         3,\core\output\notification::NOTIFY_WARNING);
             }
 
-            //first delete subs
-            //what to do about upstream???
+            //only delete if its been confirmed
+            if ($confirm and confirm_sesskey()) {
+                //what to do about upstream???
+                $result=$DB->delete_records(constants::M_TABLE_SCHOOLS,array('id'=>$id));
+                redirect($returnurl);
+            }
 
-            //
             $strheading = get_string('deleteschool', constants::M_COMP);
             $PAGE->navbar->add($strheading);
             $PAGE->set_title($strheading);
@@ -130,7 +136,7 @@ if ($delete && $id) {
             echo $renderer->heading($strheading);
             $yesurl = new moodle_url($baseurl . '/subs/edit.php', array('id' => $id, 'delete' => 1,'type'=>'school',
                     'confirm' => 1, 'sesskey' => sesskey(), 'returnurl' => $returnurl->out_as_local_url()));
-            $message = get_string('deleteschoolconfirm', constants::M_COMP);
+            $message = get_string('deleteschoolconfirm', constants::M_COMP,  $school->name);
             echo $renderer->confirm($message, $yesurl, $returnurl);
             echo $renderer->footer();
             die;
